Prepare to use fable minimal app
Previous commits began with the project structure recommended by the
author. This starting place is fine (or at least was at one time), but
this starting point causes issues. For exmaple, the version of
`fable-compiler` is 2.x but 3.x is now "standard" (and 4.x is in alpha).
Additionally, I think that perhaps my issues with Node versions and
other errors may have resulted from this starting point.